+++ src/proto/drawline.pro
@@ -1,0 +1,3 @@
+/* drawline.c */
+int win_line(win_T *wp, linenr_T lnum, int startrow, int endrow, int nochange, int number_only);
+/* vim: set ft=c : */
